UI Storyboard Progress
Just to let you guys have an idea about how's the storyboard doing. It's still not done yet: 1) There are some more arrows to draw 2) Figma somehow doesn't support exporting the diagram (one picture with all the frames and arrows) so I might have to zoom in and cut and paste and put it all together so the picture includes everything while being clear enough  (the picture I uploaded here is a screenshot btw so it's super blurry).